+++ NEWS.md
@@ -1,5 +1,6 @@
 ## pyobsplot 0.3.4dev
 
+- Add small padding around figure outputs for jsdom renderer to improve presentation over non-white backgrounds.
 - Update Observable Plot to 0.6.6.